fixed LOGFONT

+++ drawer/src/main/java/jp/chang/myclinic/drawer/printer/LOGFONT.java
@@ -1,33 +1,27 @@
 package jp.chang.myclinic.drawer.printer;
 
-import com.sun.jna.platform.win32.WinUser;
-import com.sun.jna.platform.win32.WinDef;
-import com.sun.jna.platform.win32.Kernel32;
-import com.sun.jna.platform.win32.User32;
-import com.sun.jna.win32.W32APITypeMapper;
-import com.sun.jna.Native;
 import com.sun.jna.Structure;
-import com.sun.jna.Pointer;
-import com.sun.jna.WString;
+import com.sun.jna.platform.win32.WinDef;
+import com.sun.jna.platform.win32.WinUser;
 
-import java.util.List;
 import java.util.Arrays;
+import java.util.List;
 
 public class LOGFONT extends Structure implements WinDef, WinUser {
 	
-	public long lfHeight;
-	public long lfWidth;
-	public long lfEscapement;
-	public long lfOrientation;
-	public long lfWeight;
-	public long lfItalic;
-	public long lfUnderline;
-	public long lfStrikeOut;
-	public long lfCharSet;
-	public long lfOutPrecision;
-	public long lfClipPrecision;
-	public long lfQuality;
-	public long lfPitchAndFamily;
+	public LONG lfHeight;
+	public LONG lfWidth;
+	public LONG lfEscapement;
+	public LONG lfOrientation;
+	public LONG lfWeight;
+	public BYTE lfItalic;
+	public BYTE lfUnderline;
+	public BYTE lfStrikeOut;
+	public BYTE lfCharSet;
+	public BYTE lfOutPrecision;
+	public BYTE lfClipPrecision;
+	public BYTE lfQuality;
+	public BYTE lfPitchAndFamily;
 	public char[] lfFaceName = new char[PrinterConsts.LF_FACESIZE];
 
 	@Override
